First steps of implementation
This commit provides a new architecture that is simpler
and might provide better performances. Instead of running
everything in threads, the network part is handle in the
main loop (but could be handled in another loop) while
all the heavy work (parsing information, creating models)
is done in worker threads, managed by a threadpool.