Adding automated scripts
This commit provides some automated scripts, written
in python, that should help generating code that is
boring to write.

Currently, a simple generator that is able to create
code for container classes is done. Another generator
that will be able to generate processor and loader
code will be written as well.