Converted objects API and reliable generator
This commit provides a new API for Facebook objects
as well as a reliable generator. From now, many
processors should not be needed, so the processing
task is now greately simplified.

Loaders are still needed since they are the bridge
between C++ and QML and properties cannot be exposed
dynamically in QML (they need to be described with
Q_PROPERTY).